Move API conformance testing under v1alpha1 sub-package.
