+++ test/conformance/api/v1alpha1/revision_timeout_test.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package api
+package v1alpha1
 
 import (
 	"encoding/json"
