+++ test/conformance/api/v1alpha1/workingdir_test.go
@@ -16,10 +16,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package api
+package v1alpha1
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
@@ -27,49 +26,32 @@ import (
 	v1a1test "github.com/knative/serving/test/v1alpha1"
 )
 
-func TestCmdArgsService(t *testing.T) {
+func TestWorkingDirService(t *testing.T) {
 	t.Parallel()
 	clients := test.Setup(t)
 
 	names := test.ResourceNames{
 		Service: test.ObjectNameForTest(t),
-		Image:   "python:3",
+		Image:   test.WorkingDir,
 	}
 
 	// Clean up on test failure or interrupt
 	defer test.TearDown(clients, names)
 	test.CleanupOnInterrupt(func() { test.TearDown(clients, names) })
 
-	const text = "THIS IS THE CMD AND ARGS TEST"
+	const wd = "/foo/bar/baz"
 
 	// Setup initial Service
 	_, err := v1a1test.CreateRunLatestServiceReady(t, clients, &names, &v1a1test.Options{},
 		func(svc *v1alpha1.Service) {
 			c := &svc.Spec.Template.Spec.Containers[0]
-			c.Image = names.Image
-			c.Command = []string{"python"}
-			c.Args = []string{"-c", fmt.Sprintf(`
-import http.server
-import socketserver
-from http import HTTPStatus
-
-
-class Handler(http.server.SimpleHTTPRequestHandler):
-    def do_GET(self):
-        self.send_response(HTTPStatus.OK)
-        self.end_headers()
-        self.wfile.write(b'%s')
-
-
-httpd = socketserver.TCPServer(('', 8080), Handler)
-httpd.serve_forever()`, text),
-			}
+			c.WorkingDir = wd
 		})
 	if err != nil {
 		t.Fatalf("Failed to create initial Service %v: %v", names.Service, err)
 	}
 
-	if err = validateRunLatestDataPlane(t, clients, names, text); err != nil {
+	if err = validateRunLatestDataPlane(t, clients, names, wd); err != nil {
 		t.Error(err)
 	}
 }
